Crate librocksdb_sys

Structs

__fsid_t

__va_list_tag

rocksdb_backup_engine_info_t

rocksdb_backup_engine_options_t

rocksdb_backup_engine_t

rocksdb_block_based_table_options_t

rocksdb_cache_t

rocksdb_checkpoint_t

rocksdb_column_family_handle_t

rocksdb_column_family_metadata_t

rocksdb_compactionfilter_t

rocksdb_compactionfiltercontext_t

rocksdb_compactionfilterfactory_t

rocksdb_compactoptions_t

rocksdb_comparator_t

rocksdb_cuckoo_table_options_t

rocksdb_dbpath_t

rocksdb_env_t

rocksdb_envoptions_t

rocksdb_fifo_compaction_options_t

rocksdb_filelock_t

rocksdb_filterpolicy_t

rocksdb_flushoptions_t

rocksdb_hyper_clock_cache_options_t

rocksdb_ingestexternalfileoptions_t

rocksdb_iterator_t

rocksdb_level_metadata_t

rocksdb_livefiles_t

rocksdb_logger_t

rocksdb_lru_cache_options_t

rocksdb_memory_allocator_t

rocksdb_memory_consumers_t

rocksdb_memory_usage_t

rocksdb_mergeoperator_t

rocksdb_optimistictransaction_options_t

rocksdb_optimistictransactiondb_t

rocksdb_options_t

rocksdb_perfcontext_t

rocksdb_pinnableslice_t

rocksdb_randomfile_t

rocksdb_ratelimiter_t

rocksdb_readoptions_t

rocksdb_restore_options_t

rocksdb_seqfile_t

rocksdb_slicetransform_t

rocksdb_snapshot_t

rocksdb_sst_file_metadata_t

rocksdb_sstfilewriter_t

rocksdb_t

rocksdb_transaction_options_t

rocksdb_transaction_t

rocksdb_transactiondb_options_t

rocksdb_transactiondb_t

rocksdb_universal_compaction_options_t

rocksdb_wal_iterator_t

rocksdb_wal_readoptions_t

rocksdb_writablefile_t

rocksdb_writebatch_t

rocksdb_writebatch_wi_t

rocksdb_writeoptions_t

Constants

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

PTRDIFF_MAX

PTRDIFF_MIN

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WINT_MAX

WINT_MIN

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_LEAST_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_FEATURES_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_SYS_CDEFS_H

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_C2X_STRTOL

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_IEC_60559_BFP__

__STDC_IEC_60559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__bool_true_false_are_defined

__glibc_c99_flexarr_available

false_

rocksdb_absolute_consistency_recovery

rocksdb_blob_cache_hit_count

rocksdb_blob_checksum_time

rocksdb_blob_decompress_time

rocksdb_blob_read_byte

rocksdb_blob_read_count

rocksdb_blob_read_time

rocksdb_block_based_table_data_block_index_type_binary_search

rocksdb_block_based_table_data_block_index_type_binary_search_and_hash

rocksdb_block_based_table_index_type_binary_search

rocksdb_block_based_table_index_type_hash_search

rocksdb_block_based_table_index_type_two_level_index_search

rocksdb_block_cache_hit_count

rocksdb_block_checksum_time

rocksdb_block_decompress_time

rocksdb_block_read_byte

rocksdb_block_read_count

rocksdb_block_read_time

rocksdb_block_seek_nanos

rocksdb_bloom_memtable_hit_count

rocksdb_bloom_memtable_miss_count

rocksdb_bloom_sst_hit_count

rocksdb_bloom_sst_miss_count

rocksdb_bz2_compression

rocksdb_db_condition_wait_nanos

rocksdb_db_mutex_lock_nanos

rocksdb_disable

rocksdb_enable_count

rocksdb_enable_time

rocksdb_enable_time_except_for_mutex

rocksdb_env_create_dir_if_missing_nanos

rocksdb_env_create_dir_nanos

rocksdb_env_delete_dir_nanos

rocksdb_env_delete_file_nanos

rocksdb_env_file_exists_nanos

rocksdb_env_get_children_file_attributes_nanos

rocksdb_env_get_children_nanos

rocksdb_env_get_file_modification_time_nanos

rocksdb_env_get_file_size_nanos

rocksdb_env_link_file_nanos

rocksdb_env_lock_file_nanos

rocksdb_env_new_directory_nanos

rocksdb_env_new_logger_nanos

rocksdb_env_new_random_access_file_nanos

rocksdb_env_new_random_rw_file_nanos

rocksdb_env_new_sequential_file_nanos

rocksdb_env_new_writable_file_nanos

rocksdb_env_rename_file_nanos

rocksdb_env_reuse_writable_file_nanos

rocksdb_env_unlock_file_nanos

rocksdb_fifo_compaction

rocksdb_find_next_user_entry_time

rocksdb_find_table_nanos

rocksdb_get_from_memtable_count

rocksdb_get_from_memtable_time

rocksdb_get_from_output_files_time

rocksdb_get_post_process_time

rocksdb_get_read_bytes

rocksdb_get_snapshot_time

rocksdb_internal_delete_skipped_count

rocksdb_internal_key_skipped_count

rocksdb_internal_merge_count

rocksdb_internal_range_del_reseek_count

rocksdb_internal_recent_skipped_count

rocksdb_iter_read_bytes

rocksdb_key_lock_wait_count

rocksdb_key_lock_wait_time

rocksdb_level_compaction

rocksdb_lz4_compression

rocksdb_lz4hc_compression

rocksdb_merge_operator_time_nanos

rocksdb_multiget_read_bytes

rocksdb_new_table_block_iter_nanos

rocksdb_new_table_iterator_nanos

rocksdb_next_on_memtable_count

rocksdb_no_compression

rocksdb_number_async_seek

rocksdb_out_of_bounds

rocksdb_point_in_time_recovery

rocksdb_prepopulate_blob_disable

rocksdb_prepopulate_blob_flush_only

rocksdb_prev_on_memtable_count

rocksdb_read_filter_block_nanos

rocksdb_read_index_block_nanos

rocksdb_seek_child_seek_count

rocksdb_seek_child_seek_time

rocksdb_seek_internal_seek_time

rocksdb_seek_max_heap_time

rocksdb_seek_min_heap_time

rocksdb_seek_on_memtable_count

rocksdb_seek_on_memtable_time

rocksdb_similar_size_compaction_stop_style

rocksdb_skip_any_corrupted_records_recovery

rocksdb_snappy_compression

rocksdb_tolerate_corrupted_tail_records_recovery

rocksdb_total_metric_count

rocksdb_total_size_compaction_stop_style

rocksdb_uninitialized

rocksdb_universal_compaction

rocksdb_user_key_comparison_count

rocksdb_write_delay_time

rocksdb_write_memtable_time

rocksdb_write_pre_and_post_process_time

rocksdb_write_wal_time

rocksdb_xpress_compression

rocksdb_zlib_compression

rocksdb_zstd_compression

true_

Functions

rocksdb_approximate_memory_usage_createโ

rocksdb_approximate_memory_usage_destroyโ

rocksdb_approximate_memory_usage_get_cache_totalโ

rocksdb_approximate_memory_usage_get_mem_table_readers_totalโ

rocksdb_approximate_memory_usage_get_mem_table_totalโ

rocksdb_approximate_memory_usage_get_mem_table_unflushedโ

rocksdb_approximate_sizesโ

rocksdb_approximate_sizes_cfโ

rocksdb_backup_engine_closeโ

rocksdb_backup_engine_create_new_backupโ

rocksdb_backup_engine_create_new_backup_flushโ

rocksdb_backup_engine_get_backup_infoโ

rocksdb_backup_engine_info_backup_idโ

rocksdb_backup_engine_info_countโ

rocksdb_backup_engine_info_destroyโ

rocksdb_backup_engine_info_number_filesโ

rocksdb_backup_engine_info_sizeโ

rocksdb_backup_engine_info_timestampโ

rocksdb_backup_engine_openโ

rocksdb_backup_engine_open_optsโ

rocksdb_backup_engine_options_createโ

rocksdb_backup_engine_options_destroyโ

rocksdb_backup_engine_options_get_backup_log_filesโ

rocksdb_backup_engine_options_get_backup_rate_limitโ

rocksdb_backup_engine_options_get_callback_trigger_interval_sizeโ

rocksdb_backup_engine_options_get_destroy_old_dataโ

rocksdb_backup_engine_options_get_max_background_operationsโ

rocksdb_backup_engine_options_get_max_valid_backups_to_openโ

rocksdb_backup_engine_options_get_restore_rate_limitโ

rocksdb_backup_engine_options_get_share_files_with_checksum_namingโ

rocksdb_backup_engine_options_get_share_table_filesโ

rocksdb_backup_engine_options_get_syncโ

rocksdb_backup_engine_options_set_backup_dirโ

rocksdb_backup_engine_options_set_backup_log_filesโ

rocksdb_backup_engine_options_set_backup_rate_limitโ

rocksdb_backup_engine_options_set_callback_trigger_interval_sizeโ

rocksdb_backup_engine_options_set_destroy_old_dataโ

rocksdb_backup_engine_options_set_envโ

rocksdb_backup_engine_options_set_max_background_operationsโ

rocksdb_backup_engine_options_set_max_valid_backups_to_openโ

rocksdb_backup_engine_options_set_restore_rate_limitโ

rocksdb_backup_engine_options_set_share_files_with_checksum_namingโ

rocksdb_backup_engine_options_set_share_table_filesโ

rocksdb_backup_engine_options_set_syncโ

rocksdb_backup_engine_purge_old_backupsโ

rocksdb_backup_engine_restore_db_from_backupโ

rocksdb_backup_engine_restore_db_from_latest_backupโ

rocksdb_backup_engine_verify_backupโ

rocksdb_batched_multi_get_cfโ

rocksdb_block_based_options_createโ

rocksdb_block_based_options_destroyโ

rocksdb_block_based_options_set_block_cacheโ

rocksdb_block_based_options_set_block_restart_intervalโ

rocksdb_block_based_options_set_block_sizeโ

rocksdb_block_based_options_set_block_size_deviationโ

rocksdb_block_based_options_set_cache_index_and_filter_blocksโ

rocksdb_block_based_options_set_cache_index_and_filter_blocks_with_high_priorityโ

rocksdb_block_based_options_set_checksumโ

rocksdb_block_based_options_set_data_block_hash_ratioโ

rocksdb_block_based_options_set_data_block_index_typeโ

rocksdb_block_based_options_set_filter_policyโ

rocksdb_block_based_options_set_format_versionโ

rocksdb_block_based_options_set_index_block_restart_intervalโ

rocksdb_block_based_options_set_index_typeโ

rocksdb_block_based_options_set_metadata_block_sizeโ

rocksdb_block_based_options_set_no_block_cacheโ

rocksdb_block_based_options_set_optimize_filters_for_memoryโ

rocksdb_block_based_options_set_partition_filtersโ

rocksdb_block_based_options_set_pin_l0_filter_and_index_blocks_in_cacheโ

rocksdb_block_based_options_set_pin_top_level_index_and_filterโ

rocksdb_block_based_options_set_use_delta_encodingโ

rocksdb_block_based_options_set_whole_key_filteringโ

rocksdb_cache_create_hyper_clockโ

rocksdb_cache_create_hyper_clock_optsโ

rocksdb_cache_create_lruโ

rocksdb_cache_create_lru_optsโ

rocksdb_cache_create_lru_with_strict_capacity_limitโ

rocksdb_cache_destroyโ

rocksdb_cache_disown_dataโ

rocksdb_cache_get_capacityโ

rocksdb_cache_get_pinned_usageโ

rocksdb_cache_get_usageโ

rocksdb_cache_set_capacityโ

rocksdb_cancel_all_background_workโ

rocksdb_checkpoint_createโ

rocksdb_checkpoint_object_createโ

rocksdb_checkpoint_object_destroyโ

rocksdb_closeโ

rocksdb_column_family_handle_destroyโ

rocksdb_column_family_handle_get_idโ

rocksdb_column_family_handle_get_nameโ

rocksdb_column_family_metadata_destroyโ

rocksdb_column_family_metadata_get_file_countโ

rocksdb_column_family_metadata_get_level_countโ

rocksdb_column_family_metadata_get_level_metadataโ

Returns the rocksdb_level_metadata_t of the ith level from the specified column family metadata.

rocksdb_column_family_metadata_get_nameโ

rocksdb_column_family_metadata_get_sizeโ

rocksdb_compact_rangeโ

rocksdb_compact_range_cfโ

rocksdb_compact_range_cf_optโ

rocksdb_compact_range_optโ

rocksdb_compactionfilter_createโ

rocksdb_compactionfilter_destroyโ

rocksdb_compactionfilter_set_ignore_snapshotsโ

rocksdb_compactionfiltercontext_is_full_compactionโ

rocksdb_compactionfiltercontext_is_manual_compactionโ

rocksdb_compactionfilterfactory_createโ

rocksdb_compactionfilterfactory_destroyโ

rocksdb_compactoptions_createโ

rocksdb_compactoptions_destroyโ

rocksdb_compactoptions_get_bottommost_level_compactionโ

rocksdb_compactoptions_get_change_levelโ

rocksdb_compactoptions_get_exclusive_manual_compactionโ

rocksdb_compactoptions_get_target_levelโ

rocksdb_compactoptions_set_bottommost_level_compactionโ

rocksdb_compactoptions_set_change_levelโ

rocksdb_compactoptions_set_exclusive_manual_compactionโ

rocksdb_compactoptions_set_full_history_ts_lowโ

rocksdb_compactoptions_set_target_levelโ

rocksdb_comparator_createโ

rocksdb_comparator_destroyโ

rocksdb_comparator_with_ts_createโ

rocksdb_create_column_familyโ

rocksdb_create_column_family_with_ttlโ

rocksdb_create_default_envโ

rocksdb_create_dir_if_missingโ

rocksdb_create_iteratorโ

rocksdb_create_iterator_cfโ

rocksdb_create_iteratorsโ

rocksdb_create_mem_envโ

rocksdb_create_snapshotโ

rocksdb_cuckoo_options_createโ

rocksdb_cuckoo_options_destroyโ

rocksdb_cuckoo_options_set_cuckoo_block_sizeโ

rocksdb_cuckoo_options_set_hash_ratioโ

rocksdb_cuckoo_options_set_identity_as_first_hashโ

rocksdb_cuckoo_options_set_max_search_depthโ

rocksdb_cuckoo_options_set_use_module_hashโ

rocksdb_dbpath_createโ

rocksdb_dbpath_destroyโ

rocksdb_deleteโ

rocksdb_delete_cfโ

rocksdb_delete_cf_with_tsโ

rocksdb_delete_fileโ

rocksdb_delete_file_in_rangeโ

rocksdb_delete_file_in_range_cfโ

rocksdb_delete_range_cfโ

rocksdb_delete_with_tsโ

rocksdb_destroy_dbโ

rocksdb_disable_file_deletionsโ

rocksdb_disable_manual_compactionโ

rocksdb_drop_column_familyโ

rocksdb_enable_file_deletionsโ

rocksdb_enable_manual_compactionโ

rocksdb_env_destroyโ

rocksdb_env_get_background_threadsโ

rocksdb_env_get_bottom_priority_background_threadsโ

rocksdb_env_get_high_priority_background_threadsโ

rocksdb_env_get_low_priority_background_threadsโ

rocksdb_env_join_all_threadsโ

rocksdb_env_lower_high_priority_thread_pool_cpu_priorityโ

rocksdb_env_lower_high_priority_thread_pool_io_priorityโ

rocksdb_env_lower_thread_pool_cpu_priorityโ

rocksdb_env_lower_thread_pool_io_priorityโ

rocksdb_env_set_background_threadsโ

rocksdb_env_set_bottom_priority_background_threadsโ

rocksdb_env_set_high_priority_background_threadsโ

rocksdb_env_set_low_priority_background_threadsโ

rocksdb_envoptions_createโ

rocksdb_envoptions_destroyโ

rocksdb_fifo_compaction_options_createโ

rocksdb_fifo_compaction_options_destroyโ

rocksdb_fifo_compaction_options_get_allow_compactionโ

rocksdb_fifo_compaction_options_get_max_table_files_sizeโ

rocksdb_fifo_compaction_options_set_allow_compactionโ

rocksdb_fifo_compaction_options_set_max_table_files_sizeโ

rocksdb_filterpolicy_create_bloomโ

rocksdb_filterpolicy_create_bloom_fullโ

rocksdb_filterpolicy_create_ribbonโ

rocksdb_filterpolicy_create_ribbon_hybridโ

rocksdb_filterpolicy_destroyโ

rocksdb_flushโ

rocksdb_flush_cfโ

rocksdb_flush_cfsโ

rocksdb_flush_walโ

rocksdb_flushoptions_createโ

rocksdb_flushoptions_destroyโ

rocksdb_flushoptions_get_waitโ

rocksdb_flushoptions_set_waitโ

rocksdb_freeโ

rocksdb_getโ

rocksdb_get_cfโ

rocksdb_get_cf_with_tsโ

rocksdb_get_column_family_metadataโ

rocksdb_get_column_family_metadata_cfโ

Returns the rocksdb_column_family_metadata_t of the specified column family.

rocksdb_get_full_history_ts_lowโ

rocksdb_get_latest_sequence_numberโ

rocksdb_get_options_from_stringโ

rocksdb_get_pinnedโ

rocksdb_get_pinned_cfโ

rocksdb_get_updates_sinceโ

rocksdb_get_with_tsโ

rocksdb_hyper_clock_cache_options_createโ

rocksdb_hyper_clock_cache_options_destroyโ

rocksdb_hyper_clock_cache_options_set_capacityโ

rocksdb_hyper_clock_cache_options_set_estimated_entry_chargeโ

rocksdb_hyper_clock_cache_options_set_memory_allocatorโ

rocksdb_hyper_clock_cache_options_set_num_shard_bitsโ

rocksdb_increase_full_history_ts_lowโ

rocksdb_ingest_external_fileโ

rocksdb_ingest_external_file_cfโ

rocksdb_ingestexternalfileoptions_createโ

rocksdb_ingestexternalfileoptions_destroyโ

rocksdb_ingestexternalfileoptions_set_allow_blocking_flushโ

rocksdb_ingestexternalfileoptions_set_allow_global_seqnoโ

rocksdb_ingestexternalfileoptions_set_fail_if_not_bottommost_levelโ

rocksdb_ingestexternalfileoptions_set_ingest_behindโ

rocksdb_ingestexternalfileoptions_set_move_filesโ

rocksdb_ingestexternalfileoptions_set_snapshot_consistencyโ

rocksdb_iter_destroyโ

rocksdb_iter_get_errorโ

rocksdb_iter_keyโ

rocksdb_iter_nextโ

rocksdb_iter_prevโ

rocksdb_iter_seekโ

rocksdb_iter_seek_for_prevโ

rocksdb_iter_seek_to_firstโ

rocksdb_iter_seek_to_lastโ

rocksdb_iter_timestampโ

rocksdb_iter_validโ

rocksdb_iter_valueโ

rocksdb_jemalloc_nodump_allocator_createโ

rocksdb_key_may_existโ

rocksdb_key_may_exist_cfโ

rocksdb_level_metadata_destroyโ

Releases the specified rocksdb_level_metadata_t.

rocksdb_level_metadata_get_file_countโ

rocksdb_level_metadata_get_levelโ

rocksdb_level_metadata_get_sizeโ

rocksdb_level_metadata_get_sst_file_metadataโ

Returns the sst_file_metadata_t of the ith file from the specified level metadata.

rocksdb_list_column_familiesโ

rocksdb_list_column_families_destroyโ

rocksdb_livefilesโ

rocksdb_livefiles_column_family_nameโ

rocksdb_livefiles_countโ

rocksdb_livefiles_deletionsโ

rocksdb_livefiles_destroyโ

rocksdb_livefiles_entriesโ

rocksdb_livefiles_largestkeyโ

rocksdb_livefiles_levelโ

rocksdb_livefiles_nameโ

rocksdb_livefiles_sizeโ

rocksdb_livefiles_smallestkeyโ

rocksdb_load_latest_optionsโ

rocksdb_load_latest_options_destroyโ

rocksdb_lru_cache_options_createโ

rocksdb_lru_cache_options_destroyโ

rocksdb_lru_cache_options_set_capacityโ

rocksdb_lru_cache_options_set_memory_allocatorโ

rocksdb_lru_cache_options_set_num_shard_bitsโ

rocksdb_memory_allocator_destroyโ

rocksdb_memory_consumers_add_cacheโ

rocksdb_memory_consumers_add_dbโ

rocksdb_memory_consumers_createโ

rocksdb_memory_consumers_destroyโ

rocksdb_mergeโ

rocksdb_merge_cfโ

rocksdb_mergeoperator_createโ

rocksdb_mergeoperator_destroyโ

rocksdb_multi_getโ

rocksdb_multi_get_cfโ

rocksdb_multi_get_cf_with_tsโ

rocksdb_multi_get_with_tsโ

rocksdb_openโ

rocksdb_open_and_trim_historyโ

rocksdb_open_as_secondaryโ

rocksdb_open_as_secondary_column_familiesโ

rocksdb_open_column_familiesโ

rocksdb_open_column_families_with_ttlโ

rocksdb_open_for_read_onlyโ

rocksdb_open_for_read_only_column_familiesโ

rocksdb_open_with_ttlโ

rocksdb_optimistictransaction_beginโ

rocksdb_optimistictransaction_options_createโ

rocksdb_optimistictransaction_options_destroyโ

rocksdb_optimistictransaction_options_set_set_snapshotโ

rocksdb_optimistictransactiondb_checkpoint_object_createโ

rocksdb_optimistictransactiondb_closeโ

rocksdb_optimistictransactiondb_close_base_dbโ

rocksdb_optimistictransactiondb_get_base_dbโ

rocksdb_optimistictransactiondb_openโ

rocksdb_optimistictransactiondb_open_column_familiesโ

rocksdb_optimistictransactiondb_property_intโ

rocksdb_optimistictransactiondb_property_valueโ

rocksdb_optimistictransactiondb_writeโ

rocksdb_options_add_compact_on_deletion_collector_factoryโ

rocksdb_options_compaction_readahead_sizeโ

rocksdb_options_createโ

rocksdb_options_create_copyโ

rocksdb_options_destroyโ

rocksdb_options_enable_statisticsโ

rocksdb_options_get_WAL_size_limit_MBโ

rocksdb_options_get_WAL_ttl_secondsโ

rocksdb_options_get_access_hint_on_compaction_startโ

rocksdb_options_get_advise_random_on_openโ

rocksdb_options_get_allow_concurrent_memtable_writeโ

rocksdb_options_get_allow_ingest_behindโ

rocksdb_options_get_allow_mmap_readsโ

rocksdb_options_get_allow_mmap_writesโ

rocksdb_options_get_arena_block_sizeโ

rocksdb_options_get_atomic_flushโ

rocksdb_options_get_avoid_unnecessary_blocking_ioโ

rocksdb_options_get_blob_compaction_readahead_sizeโ

rocksdb_options_get_blob_compression_typeโ

rocksdb_options_get_blob_file_sizeโ

rocksdb_options_get_blob_file_starting_levelโ

rocksdb_options_get_blob_gc_age_cutoffโ

rocksdb_options_get_blob_gc_force_thresholdโ

rocksdb_options_get_bloom_localityโ

rocksdb_options_get_bottommost_compressionโ

rocksdb_options_get_bottommost_compression_options_use_zstd_dict_trainerโ

rocksdb_options_get_bytes_per_syncโ

rocksdb_options_get_compaction_readahead_sizeโ

rocksdb_options_get_compaction_styleโ

rocksdb_options_get_compressionโ

rocksdb_options_get_compression_options_max_dict_buffer_bytesโ

rocksdb_options_get_compression_options_parallel_threadsโ

rocksdb_options_get_compression_options_use_zstd_dict_trainerโ

rocksdb_options_get_compression_options_zstd_max_train_bytesโ

rocksdb_options_get_create_if_missingโ

rocksdb_options_get_create_missing_column_familiesโ

rocksdb_options_get_db_write_buffer_sizeโ

rocksdb_options_get_delete_obsolete_files_period_microsโ

rocksdb_options_get_disable_auto_compactionsโ

rocksdb_options_get_enable_blob_filesโ

rocksdb_options_get_enable_blob_gcโ

rocksdb_options_get_enable_pipelined_writeโ

rocksdb_options_get_enable_write_thread_adaptive_yieldโ

rocksdb_options_get_error_if_existsโ

rocksdb_options_get_experimental_mempurge_thresholdโ

rocksdb_options_get_hard_pending_compaction_bytes_limitโ

rocksdb_options_get_info_log_levelโ

rocksdb_options_get_inplace_update_num_locksโ

rocksdb_options_get_inplace_update_supportโ

rocksdb_options_get_is_fd_close_on_execโ

rocksdb_options_get_keep_log_file_numโ

rocksdb_options_get_level0_file_num_compaction_triggerโ

rocksdb_options_get_level0_slowdown_writes_triggerโ

rocksdb_options_get_level0_stop_writes_triggerโ

rocksdb_options_get_level_compaction_dynamic_level_bytesโ

rocksdb_options_get_log_file_time_to_rollโ

rocksdb_options_get_manifest_preallocation_sizeโ

rocksdb_options_get_manual_wal_flushโ

rocksdb_options_get_max_background_compactionsโ

rocksdb_options_get_max_background_flushesโ

rocksdb_options_get_max_background_jobsโ

rocksdb_options_get_max_bytes_for_level_baseโ

rocksdb_options_get_max_bytes_for_level_multiplierโ

rocksdb_options_get_max_compaction_bytesโ

rocksdb_options_get_max_file_opening_threadsโ

rocksdb_options_get_max_log_file_sizeโ

rocksdb_options_get_max_manifest_file_sizeโ

rocksdb_options_get_max_open_filesโ

rocksdb_options_get_max_sequential_skip_in_iterationsโ

rocksdb_options_get_max_subcompactionsโ

rocksdb_options_get_max_successive_mergesโ

rocksdb_options_get_max_total_wal_sizeโ

rocksdb_options_get_max_write_buffer_numberโ

rocksdb_options_get_max_write_buffer_number_to_maintainโ

rocksdb_options_get_max_write_buffer_size_to_maintainโ

rocksdb_options_get_memtable_huge_page_sizeโ

rocksdb_options_get_memtable_prefix_bloom_size_ratioโ

rocksdb_options_get_min_blob_sizeโ

rocksdb_options_get_min_write_buffer_number_to_mergeโ

rocksdb_options_get_num_levelsโ

rocksdb_options_get_optimize_filters_for_hitsโ

rocksdb_options_get_paranoid_checksโ

rocksdb_options_get_prepopulate_blob_cacheโ

rocksdb_options_get_recycle_log_file_numโ

rocksdb_options_get_report_bg_io_statsโ

rocksdb_options_get_skip_checking_sst_file_sizes_on_db_openโ

rocksdb_options_get_skip_stats_update_on_db_openโ

rocksdb_options_get_soft_pending_compaction_bytes_limitโ

rocksdb_options_get_stats_dump_period_secโ

rocksdb_options_get_stats_persist_period_secโ

rocksdb_options_get_table_cache_numshardbitsโ

rocksdb_options_get_target_file_size_baseโ

rocksdb_options_get_target_file_size_multiplierโ

rocksdb_options_get_unordered_writeโ

rocksdb_options_get_use_adaptive_mutexโ

rocksdb_options_get_use_direct_io_for_flush_and_compactionโ

rocksdb_options_get_use_direct_readsโ

rocksdb_options_get_use_fsyncโ

rocksdb_options_get_wal_bytes_per_syncโ

rocksdb_options_get_wal_compressionโ

rocksdb_options_get_wal_recovery_modeโ

rocksdb_options_get_writable_file_max_buffer_sizeโ

rocksdb_options_get_write_buffer_sizeโ

rocksdb_options_increase_parallelismโ

rocksdb_options_optimize_for_point_lookupโ

rocksdb_options_optimize_level_style_compactionโ

rocksdb_options_optimize_universal_style_compactionโ

rocksdb_options_prepare_for_bulk_loadโ

rocksdb_options_set_WAL_size_limit_MBโ

rocksdb_options_set_WAL_ttl_secondsโ

rocksdb_options_set_access_hint_on_compaction_startโ

rocksdb_options_set_advise_random_on_openโ

rocksdb_options_set_allow_concurrent_memtable_writeโ

rocksdb_options_set_allow_ingest_behindโ

rocksdb_options_set_allow_mmap_readsโ

rocksdb_options_set_allow_mmap_writesโ

rocksdb_options_set_arena_block_sizeโ

rocksdb_options_set_atomic_flushโ

rocksdb_options_set_avoid_unnecessary_blocking_ioโ

rocksdb_options_set_blob_cacheโ

rocksdb_options_set_blob_compaction_readahead_sizeโ

rocksdb_options_set_blob_compression_typeโ

rocksdb_options_set_blob_file_sizeโ

rocksdb_options_set_blob_file_starting_levelโ

rocksdb_options_set_blob_gc_age_cutoffโ

rocksdb_options_set_blob_gc_force_thresholdโ

rocksdb_options_set_block_based_table_factoryโ

rocksdb_options_set_bloom_localityโ

rocksdb_options_set_bottommost_compressionโ

rocksdb_options_set_bottommost_compression_optionsโ

rocksdb_options_set_bottommost_compression_options_max_dict_buffer_bytesโ

rocksdb_options_set_bottommost_compression_options_use_zstd_dict_trainerโ

rocksdb_options_set_bottommost_compression_options_zstd_max_train_bytesโ

rocksdb_options_set_bytes_per_syncโ

rocksdb_options_set_compaction_filterโ

rocksdb_options_set_compaction_filter_factoryโ

rocksdb_options_set_compaction_styleโ

rocksdb_options_set_comparatorโ

rocksdb_options_set_compressionโ

rocksdb_options_set_compression_optionsโ

rocksdb_options_set_compression_options_max_dict_buffer_bytesโ

rocksdb_options_set_compression_options_parallel_threadsโ

rocksdb_options_set_compression_options_use_zstd_dict_trainerโ

rocksdb_options_set_compression_options_zstd_max_train_bytesโ

rocksdb_options_set_compression_per_levelโ

rocksdb_options_set_create_if_missingโ

rocksdb_options_set_create_missing_column_familiesโ

rocksdb_options_set_cuckoo_table_factoryโ

rocksdb_options_set_db_log_dirโ

rocksdb_options_set_db_pathsโ

rocksdb_options_set_db_write_buffer_sizeโ

rocksdb_options_set_delete_obsolete_files_period_microsโ

rocksdb_options_set_disable_auto_compactionsโ

rocksdb_options_set_dump_malloc_statsโ

rocksdb_options_set_enable_blob_filesโ

rocksdb_options_set_enable_blob_gcโ

rocksdb_options_set_enable_pipelined_writeโ

rocksdb_options_set_enable_write_thread_adaptive_yieldโ

rocksdb_options_set_envโ

rocksdb_options_set_error_if_existsโ

rocksdb_options_set_experimental_mempurge_thresholdโ

rocksdb_options_set_fifo_compaction_optionsโ

rocksdb_options_set_hard_pending_compaction_bytes_limitโ

rocksdb_options_set_hash_link_list_repโ

rocksdb_options_set_hash_skip_list_repโ

rocksdb_options_set_info_logโ

rocksdb_options_set_info_log_levelโ

rocksdb_options_set_inplace_update_num_locksโ

rocksdb_options_set_inplace_update_supportโ

rocksdb_options_set_is_fd_close_on_execโ

rocksdb_options_set_keep_log_file_numโ

rocksdb_options_set_level0_file_num_compaction_triggerโ

rocksdb_options_set_level0_slowdown_writes_triggerโ

rocksdb_options_set_level0_stop_writes_triggerโ

rocksdb_options_set_level_compaction_dynamic_level_bytesโ

rocksdb_options_set_log_file_time_to_rollโ

rocksdb_options_set_manifest_preallocation_sizeโ

rocksdb_options_set_manual_wal_flushโ

rocksdb_options_set_max_background_compactionsโ

rocksdb_options_set_max_background_flushesโ

rocksdb_options_set_max_background_jobsโ

rocksdb_options_set_max_bytes_for_level_baseโ

rocksdb_options_set_max_bytes_for_level_multiplierโ

rocksdb_options_set_max_bytes_for_level_multiplier_additionalโ

rocksdb_options_set_max_compaction_bytesโ

rocksdb_options_set_max_file_opening_threadsโ

rocksdb_options_set_max_log_file_sizeโ

rocksdb_options_set_max_manifest_file_sizeโ

rocksdb_options_set_max_open_filesโ

rocksdb_options_set_max_sequential_skip_in_iterationsโ

rocksdb_options_set_max_subcompactionsโ

rocksdb_options_set_max_successive_mergesโ

rocksdb_options_set_max_total_wal_sizeโ

rocksdb_options_set_max_write_buffer_numberโ

rocksdb_options_set_max_write_buffer_number_to_maintainโ

rocksdb_options_set_max_write_buffer_size_to_maintainโ

rocksdb_options_set_memtable_huge_page_sizeโ

rocksdb_options_set_memtable_prefix_bloom_size_ratioโ

rocksdb_options_set_memtable_vector_repโ

rocksdb_options_set_memtable_whole_key_filteringโ

rocksdb_options_set_merge_operatorโ

rocksdb_options_set_min_blob_sizeโ

rocksdb_options_set_min_level_to_compressโ

rocksdb_options_set_min_write_buffer_number_to_mergeโ

rocksdb_options_set_num_levelsโ

rocksdb_options_set_optimize_filters_for_hitsโ

rocksdb_options_set_paranoid_checksโ

rocksdb_options_set_plain_table_factoryโ

rocksdb_options_set_prefix_extractorโ

rocksdb_options_set_prepopulate_blob_cacheโ

rocksdb_options_set_ratelimiterโ

rocksdb_options_set_recycle_log_file_numโ

rocksdb_options_set_report_bg_io_statsโ

rocksdb_options_set_row_cacheโ

rocksdb_options_set_skip_checking_sst_file_sizes_on_db_openโ

rocksdb_options_set_skip_stats_update_on_db_openโ

rocksdb_options_set_soft_pending_compaction_bytes_limitโ

rocksdb_options_set_stats_dump_period_secโ

rocksdb_options_set_stats_persist_period_secโ

rocksdb_options_set_table_cache_numshardbitsโ

rocksdb_options_set_target_file_size_baseโ

rocksdb_options_set_target_file_size_multiplierโ

rocksdb_options_set_uint64add_merge_operatorโ

rocksdb_options_set_universal_compaction_optionsโ

rocksdb_options_set_unordered_writeโ

rocksdb_options_set_use_adaptive_mutexโ

rocksdb_options_set_use_direct_io_for_flush_and_compactionโ

rocksdb_options_set_use_direct_readsโ

rocksdb_options_set_use_fsyncโ

rocksdb_options_set_wal_bytes_per_syncโ

rocksdb_options_set_wal_compressionโ

rocksdb_options_set_wal_dirโ

rocksdb_options_set_wal_recovery_modeโ

rocksdb_options_set_writable_file_max_buffer_sizeโ

rocksdb_options_set_write_buffer_sizeโ

rocksdb_options_statistics_get_stringโ

rocksdb_perfcontext_createโ

rocksdb_perfcontext_destroyโ

rocksdb_perfcontext_metricโ

rocksdb_perfcontext_reportโ

rocksdb_perfcontext_resetโ

rocksdb_pinnableslice_destroyโ

rocksdb_pinnableslice_valueโ

rocksdb_property_intโ

rocksdb_property_int_cfโ

rocksdb_property_valueโ

rocksdb_property_value_cfโ

rocksdb_putโ

rocksdb_put_cfโ

rocksdb_put_cf_with_tsโ

rocksdb_put_with_tsโ

rocksdb_ratelimiter_createโ

rocksdb_ratelimiter_destroyโ

rocksdb_readoptions_createโ

rocksdb_readoptions_destroyโ

rocksdb_readoptions_get_async_ioโ

rocksdb_readoptions_get_background_purge_on_iterator_cleanupโ

rocksdb_readoptions_get_deadlineโ

rocksdb_readoptions_get_fill_cacheโ

rocksdb_readoptions_get_ignore_range_deletionsโ

rocksdb_readoptions_get_io_timeoutโ

rocksdb_readoptions_get_max_skippable_internal_keysโ

rocksdb_readoptions_get_pin_dataโ

rocksdb_readoptions_get_prefix_same_as_startโ

rocksdb_readoptions_get_read_tierโ

rocksdb_readoptions_get_readahead_sizeโ

rocksdb_readoptions_get_tailingโ

rocksdb_readoptions_get_total_order_seekโ

rocksdb_readoptions_get_verify_checksumsโ

rocksdb_readoptions_set_async_ioโ

rocksdb_readoptions_set_background_purge_on_iterator_cleanupโ

rocksdb_readoptions_set_deadlineโ

rocksdb_readoptions_set_fill_cacheโ

rocksdb_readoptions_set_ignore_range_deletionsโ

rocksdb_readoptions_set_io_timeoutโ

rocksdb_readoptions_set_iter_start_tsโ

rocksdb_readoptions_set_iterate_lower_boundโ

rocksdb_readoptions_set_iterate_upper_boundโ

rocksdb_readoptions_set_managedโ

rocksdb_readoptions_set_max_skippable_internal_keysโ

rocksdb_readoptions_set_pin_dataโ

rocksdb_readoptions_set_prefix_same_as_startโ

rocksdb_readoptions_set_read_tierโ

rocksdb_readoptions_set_readahead_sizeโ

rocksdb_readoptions_set_snapshotโ

rocksdb_readoptions_set_tailingโ

rocksdb_readoptions_set_timestampโ

rocksdb_readoptions_set_total_order_seekโ

rocksdb_readoptions_set_verify_checksumsโ

rocksdb_release_snapshotโ

rocksdb_repair_dbโ

rocksdb_restore_options_createโ

rocksdb_restore_options_destroyโ

rocksdb_restore_options_set_keep_log_filesโ

rocksdb_set_optionsโ

rocksdb_set_options_cfโ

rocksdb_set_perf_levelโ

rocksdb_singledeleteโ

rocksdb_singledelete_cfโ

rocksdb_singledelete_cf_with_tsโ

rocksdb_singledelete_with_tsโ

rocksdb_slicetransform_createโ

rocksdb_slicetransform_create_fixed_prefixโ

rocksdb_slicetransform_create_noopโ

rocksdb_slicetransform_destroyโ

rocksdb_sst_file_metadata_destroyโ

Releases the specified rocksdb_sst_file_metadata_t.

rocksdb_sst_file_metadata_get_largestkeyโ

Returns the smallest key of the specified sst file. The caller is responsible for releasing the returned memory.

rocksdb_sst_file_metadata_get_relative_filenameโ

rocksdb_sst_file_metadata_get_sizeโ

rocksdb_sst_file_metadata_get_smallestkeyโ

Returns the smallest key of the specified sst file. The caller is responsible for releasing the returned memory.

rocksdb_sstfilewriter_addโ

rocksdb_sstfilewriter_createโ

rocksdb_sstfilewriter_create_with_comparatorโ

rocksdb_sstfilewriter_deleteโ

rocksdb_sstfilewriter_delete_rangeโ

rocksdb_sstfilewriter_delete_with_tsโ

rocksdb_sstfilewriter_destroyโ

rocksdb_sstfilewriter_file_sizeโ

rocksdb_sstfilewriter_finishโ

rocksdb_sstfilewriter_mergeโ

rocksdb_sstfilewriter_openโ

rocksdb_sstfilewriter_putโ

rocksdb_sstfilewriter_put_with_tsโ

rocksdb_suggest_compact_rangeโ

rocksdb_suggest_compact_range_cfโ

rocksdb_transaction_beginโ

rocksdb_transaction_commitโ

rocksdb_transaction_create_iteratorโ

rocksdb_transaction_create_iterator_cfโ

rocksdb_transaction_deleteโ

rocksdb_transaction_delete_cfโ

rocksdb_transaction_destroyโ

rocksdb_transaction_getโ

rocksdb_transaction_get_cfโ

rocksdb_transaction_get_for_updateโ

rocksdb_transaction_get_for_update_cfโ

rocksdb_transaction_get_nameโ

rocksdb_transaction_get_pinnedโ

rocksdb_transaction_get_pinned_cfโ

rocksdb_transaction_get_pinned_for_updateโ

rocksdb_transaction_get_pinned_for_update_cfโ

rocksdb_transaction_get_snapshotโ

rocksdb_transaction_get_writebatch_wiโ

rocksdb_transaction_mergeโ

rocksdb_transaction_merge_cfโ

rocksdb_transaction_multi_getโ

rocksdb_transaction_multi_get_cfโ

rocksdb_transaction_options_createโ

rocksdb_transaction_options_destroyโ

rocksdb_transaction_options_set_deadlock_detectโ

rocksdb_transaction_options_set_deadlock_detect_depthโ

rocksdb_transaction_options_set_expirationโ

rocksdb_transaction_options_set_lock_timeoutโ

rocksdb_transaction_options_set_max_write_batch_sizeโ

rocksdb_transaction_options_set_set_snapshotโ

rocksdb_transaction_options_set_skip_prepareโ

rocksdb_transaction_prepareโ

rocksdb_transaction_putโ

rocksdb_transaction_put_cfโ

rocksdb_transaction_rebuild_from_writebatchโ

rocksdb_transaction_rebuild_from_writebatch_wiโ

rocksdb_transaction_rollbackโ

rocksdb_transaction_rollback_to_savepointโ

rocksdb_transaction_set_commit_timestampโ

rocksdb_transaction_set_nameโ

rocksdb_transaction_set_read_timestamp_for_validationโ

rocksdb_transaction_set_savepointโ

rocksdb_transactiondb_checkpoint_object_createโ

rocksdb_transactiondb_closeโ

rocksdb_transactiondb_create_column_familyโ

rocksdb_transactiondb_create_iteratorโ

rocksdb_transactiondb_create_iterator_cfโ

rocksdb_transactiondb_create_snapshotโ

rocksdb_transactiondb_deleteโ

rocksdb_transactiondb_delete_cfโ

rocksdb_transactiondb_flushโ

rocksdb_transactiondb_flush_cfโ

rocksdb_transactiondb_flush_cfsโ

rocksdb_transactiondb_flush_walโ

rocksdb_transactiondb_getโ

rocksdb_transactiondb_get_cfโ

rocksdb_transactiondb_get_pinnedโ

rocksdb_transactiondb_get_pinned_cfโ

rocksdb_transactiondb_get_prepared_transactionsโ

rocksdb_transactiondb_mergeโ

rocksdb_transactiondb_merge_cfโ

rocksdb_transactiondb_multi_getโ

rocksdb_transactiondb_multi_get_cfโ

rocksdb_transactiondb_openโ

rocksdb_transactiondb_open_column_familiesโ

rocksdb_transactiondb_options_createโ

rocksdb_transactiondb_options_destroyโ

rocksdb_transactiondb_options_set_default_lock_timeoutโ

rocksdb_transactiondb_options_set_max_num_locksโ

rocksdb_transactiondb_options_set_num_stripesโ

rocksdb_transactiondb_options_set_transaction_lock_timeoutโ

rocksdb_transactiondb_property_intโ

rocksdb_transactiondb_property_valueโ

rocksdb_transactiondb_putโ

rocksdb_transactiondb_put_cfโ

rocksdb_transactiondb_release_snapshotโ

rocksdb_transactiondb_writeโ

rocksdb_try_catch_up_with_primaryโ

rocksdb_universal_compaction_options_createโ

rocksdb_universal_compaction_options_destroyโ

rocksdb_universal_compaction_options_get_compression_size_percentโ

rocksdb_universal_compaction_options_get_max_merge_widthโ

rocksdb_universal_compaction_options_get_max_size_amplification_percentโ

rocksdb_universal_compaction_options_get_min_merge_widthโ

rocksdb_universal_compaction_options_get_size_ratioโ

rocksdb_universal_compaction_options_get_stop_styleโ

rocksdb_universal_compaction_options_set_compression_size_percentโ

rocksdb_universal_compaction_options_set_max_merge_widthโ

rocksdb_universal_compaction_options_set_max_size_amplification_percentโ

rocksdb_universal_compaction_options_set_min_merge_widthโ

rocksdb_universal_compaction_options_set_size_ratioโ

rocksdb_universal_compaction_options_set_stop_styleโ

rocksdb_wal_iter_destroyโ

rocksdb_wal_iter_get_batchโ

rocksdb_wal_iter_nextโ

rocksdb_wal_iter_statusโ

rocksdb_wal_iter_validโ

rocksdb_writeโ

rocksdb_write_writebatch_wiโ

rocksdb_writebatch_clearโ

rocksdb_writebatch_countโ

rocksdb_writebatch_createโ

rocksdb_writebatch_create_fromโ

rocksdb_writebatch_dataโ

rocksdb_writebatch_deleteโ

rocksdb_writebatch_delete_cfโ

rocksdb_writebatch_delete_cf_with_tsโ

rocksdb_writebatch_delete_rangeโ

rocksdb_writebatch_delete_range_cfโ

rocksdb_writebatch_delete_rangevโ

rocksdb_writebatch_delete_rangev_cfโ

rocksdb_writebatch_deletevโ

rocksdb_writebatch_deletev_cfโ

rocksdb_writebatch_destroyโ

rocksdb_writebatch_iterateโ

rocksdb_writebatch_mergeโ

rocksdb_writebatch_merge_cfโ

rocksdb_writebatch_mergevโ

rocksdb_writebatch_mergev_cfโ

rocksdb_writebatch_pop_save_pointโ

rocksdb_writebatch_putโ

rocksdb_writebatch_put_cfโ

rocksdb_writebatch_put_cf_with_tsโ

rocksdb_writebatch_put_log_dataโ

rocksdb_writebatch_putvโ

rocksdb_writebatch_putv_cfโ

rocksdb_writebatch_rollback_to_save_pointโ

rocksdb_writebatch_set_save_pointโ

rocksdb_writebatch_singledeleteโ

rocksdb_writebatch_singledelete_cfโ

rocksdb_writebatch_singledelete_cf_with_tsโ

rocksdb_writebatch_wi_clearโ

rocksdb_writebatch_wi_countโ

rocksdb_writebatch_wi_createโ

rocksdb_writebatch_wi_create_fromโ

rocksdb_writebatch_wi_create_iterator_with_baseโ

rocksdb_writebatch_wi_create_iterator_with_base_cfโ

rocksdb_writebatch_wi_dataโ

rocksdb_writebatch_wi_deleteโ

rocksdb_writebatch_wi_delete_cfโ

rocksdb_writebatch_wi_delete_rangeโ

rocksdb_writebatch_wi_delete_range_cfโ

rocksdb_writebatch_wi_delete_rangevโ

rocksdb_writebatch_wi_delete_rangev_cfโ

rocksdb_writebatch_wi_deletevโ

rocksdb_writebatch_wi_deletev_cfโ

rocksdb_writebatch_wi_destroyโ

rocksdb_writebatch_wi_get_from_batchโ

rocksdb_writebatch_wi_get_from_batch_and_dbโ

rocksdb_writebatch_wi_get_from_batch_and_db_cfโ

rocksdb_writebatch_wi_get_from_batch_cfโ

rocksdb_writebatch_wi_iterateโ

rocksdb_writebatch_wi_mergeโ

rocksdb_writebatch_wi_merge_cfโ

rocksdb_writebatch_wi_mergevโ

rocksdb_writebatch_wi_mergev_cfโ

rocksdb_writebatch_wi_putโ

rocksdb_writebatch_wi_put_cfโ

rocksdb_writebatch_wi_put_log_dataโ

rocksdb_writebatch_wi_putvโ

rocksdb_writebatch_wi_putv_cfโ

rocksdb_writebatch_wi_rollback_to_save_pointโ

rocksdb_writebatch_wi_set_save_pointโ

rocksdb_writebatch_wi_singledeleteโ

rocksdb_writebatch_wi_singledelete_cfโ

rocksdb_writeoptions_createโ

rocksdb_writeoptions_destroyโ

rocksdb_writeoptions_disable_WALโ

rocksdb_writeoptions_get_disable_WALโ

rocksdb_writeoptions_get_ignore_missing_column_familiesโ

rocksdb_writeoptions_get_low_priโ

rocksdb_writeoptions_get_memtable_insert_hint_per_batchโ

rocksdb_writeoptions_get_no_slowdownโ

rocksdb_writeoptions_get_syncโ

rocksdb_writeoptions_set_ignore_missing_column_familiesโ

rocksdb_writeoptions_set_low_priโ

rocksdb_writeoptions_set_memtable_insert_hint_per_batchโ

rocksdb_writeoptions_set_no_slowdownโ

rocksdb_writeoptions_set_syncโ

Type Aliases

__blkcnt64_t

__blkcnt_t

__blksize_t

__builtin_va_list

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__gnuc_va_list

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

_bindgen_ty_1

_bindgen_ty_2

_bindgen_ty_3

_bindgen_ty_4

_bindgen_ty_5

_bindgen_ty_6

_bindgen_ty_7

_bindgen_ty_8

_bindgen_ty_9

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

va_list

wchar_t