Crate librocksdb_sys

Structs

__fsid_t

__va_list_tag

rocksdb_backup_engine_info_t

rocksdb_backup_engine_options_t

rocksdb_backup_engine_t

rocksdb_block_based_table_options_t

rocksdb_cache_t

rocksdb_checkpoint_t

rocksdb_column_family_handle_t

rocksdb_column_family_metadata_t

rocksdb_compactionfilter_t

rocksdb_compactionfiltercontext_t

rocksdb_compactionfilterfactory_t

rocksdb_compactoptions_t

rocksdb_comparator_t

rocksdb_cuckoo_table_options_t

rocksdb_dbpath_t

rocksdb_env_t

rocksdb_envoptions_t

rocksdb_fifo_compaction_options_t

rocksdb_filelock_t

rocksdb_filterpolicy_t

rocksdb_flushoptions_t

rocksdb_hyper_clock_cache_options_t

rocksdb_ingestexternalfileoptions_t

rocksdb_iterator_t

rocksdb_level_metadata_t

rocksdb_livefiles_t

rocksdb_logger_t

rocksdb_lru_cache_options_t

rocksdb_memory_allocator_t

rocksdb_memory_consumers_t

rocksdb_memory_usage_t

rocksdb_mergeoperator_t

rocksdb_optimistictransaction_options_t

rocksdb_optimistictransactiondb_t

rocksdb_options_t

rocksdb_perfcontext_t

rocksdb_pinnableslice_t

rocksdb_randomfile_t

rocksdb_ratelimiter_t

rocksdb_readoptions_t

rocksdb_restore_options_t

rocksdb_seqfile_t

rocksdb_slicetransform_t

rocksdb_snapshot_t

rocksdb_sst_file_metadata_t

rocksdb_sstfilewriter_t

rocksdb_t

rocksdb_transaction_options_t

rocksdb_transaction_t

rocksdb_transactiondb_options_t

rocksdb_transactiondb_t

rocksdb_universal_compaction_options_t

rocksdb_wal_iterator_t

rocksdb_wal_readoptions_t

rocksdb_writablefile_t

rocksdb_writebatch_t

rocksdb_writebatch_wi_t

rocksdb_writeoptions_t

Constants

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

PTRDIFF_MAX

PTRDIFF_MIN

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WINT_MAX

WINT_MIN

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_LEAST_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_FEATURES_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_SYS_CDEFS_H

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_C2X_STRTOL

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_IEC_60559_BFP__

__STDC_IEC_60559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__bool_true_false_are_defined

__glibc_c99_flexarr_available

false_

rocksdb_absolute_consistency_recovery

rocksdb_blob_cache_hit_count

rocksdb_blob_checksum_time

rocksdb_blob_decompress_time

rocksdb_blob_read_byte

rocksdb_blob_read_count

rocksdb_blob_read_time

rocksdb_block_based_table_data_block_index_type_binary_search

rocksdb_block_based_table_data_block_index_type_binary_search_and_hash

rocksdb_block_based_table_index_type_binary_search

rocksdb_block_based_table_index_type_hash_search

rocksdb_block_based_table_index_type_two_level_index_search

rocksdb_block_cache_hit_count

rocksdb_block_checksum_time

rocksdb_block_decompress_time

rocksdb_block_read_byte

rocksdb_block_read_count

rocksdb_block_read_time

rocksdb_block_seek_nanos

rocksdb_bloom_memtable_hit_count

rocksdb_bloom_memtable_miss_count

rocksdb_bloom_sst_hit_count

rocksdb_bloom_sst_miss_count

rocksdb_bz2_compression

rocksdb_db_condition_wait_nanos

rocksdb_db_mutex_lock_nanos

rocksdb_disable

rocksdb_enable_count

rocksdb_enable_time

rocksdb_enable_time_except_for_mutex

rocksdb_env_create_dir_if_missing_nanos

rocksdb_env_create_dir_nanos

rocksdb_env_delete_dir_nanos

rocksdb_env_delete_file_nanos

rocksdb_env_file_exists_nanos

rocksdb_env_get_children_file_attributes_nanos

rocksdb_env_get_children_nanos

rocksdb_env_get_file_modification_time_nanos

rocksdb_env_get_file_size_nanos

rocksdb_env_link_file_nanos

rocksdb_env_lock_file_nanos

rocksdb_env_new_directory_nanos

rocksdb_env_new_logger_nanos

rocksdb_env_new_random_access_file_nanos

rocksdb_env_new_random_rw_file_nanos

rocksdb_env_new_sequential_file_nanos

rocksdb_env_new_writable_file_nanos

rocksdb_env_rename_file_nanos

rocksdb_env_reuse_writable_file_nanos

rocksdb_env_unlock_file_nanos

rocksdb_fifo_compaction

rocksdb_find_next_user_entry_time

rocksdb_find_table_nanos

rocksdb_get_from_memtable_count

rocksdb_get_from_memtable_time

rocksdb_get_from_output_files_time

rocksdb_get_post_process_time

rocksdb_get_read_bytes

rocksdb_get_snapshot_time

rocksdb_internal_delete_skipped_count

rocksdb_internal_key_skipped_count

rocksdb_internal_merge_count

rocksdb_internal_range_del_reseek_count

rocksdb_internal_recent_skipped_count

rocksdb_iter_read_bytes

rocksdb_key_lock_wait_count

rocksdb_key_lock_wait_time

rocksdb_level_compaction

rocksdb_lz4_compression

rocksdb_lz4hc_compression

rocksdb_merge_operator_time_nanos

rocksdb_multiget_read_bytes

rocksdb_new_table_block_iter_nanos

rocksdb_new_table_iterator_nanos

rocksdb_next_on_memtable_count

rocksdb_no_compression

rocksdb_number_async_seek

rocksdb_out_of_bounds

rocksdb_point_in_time_recovery

rocksdb_prepopulate_blob_disable

rocksdb_prepopulate_blob_flush_only

rocksdb_prev_on_memtable_count

rocksdb_read_filter_block_nanos

rocksdb_read_index_block_nanos

rocksdb_seek_child_seek_count

rocksdb_seek_child_seek_time

rocksdb_seek_internal_seek_time

rocksdb_seek_max_heap_time

rocksdb_seek_min_heap_time

rocksdb_seek_on_memtable_count

rocksdb_seek_on_memtable_time

rocksdb_similar_size_compaction_stop_style

rocksdb_skip_any_corrupted_records_recovery

rocksdb_snappy_compression

rocksdb_tolerate_corrupted_tail_records_recovery

rocksdb_total_metric_count

rocksdb_total_size_compaction_stop_style

rocksdb_uninitialized

rocksdb_universal_compaction

rocksdb_user_key_comparison_count

rocksdb_write_delay_time

rocksdb_write_memtable_time

rocksdb_write_pre_and_post_process_time

rocksdb_write_wal_time

rocksdb_xpress_compression

rocksdb_zlib_compression

rocksdb_zstd_compression

true_

Functions

rocksdb_approximate_memory_usage_create⚠

rocksdb_approximate_memory_usage_destroy⚠

rocksdb_approximate_memory_usage_get_cache_total⚠

rocksdb_approximate_memory_usage_get_mem_table_readers_total⚠

rocksdb_approximate_memory_usage_get_mem_table_total⚠

rocksdb_approximate_memory_usage_get_mem_table_unflushed⚠

rocksdb_approximate_sizes⚠

rocksdb_approximate_sizes_cf⚠

rocksdb_backup_engine_close⚠

rocksdb_backup_engine_create_new_backup⚠

rocksdb_backup_engine_create_new_backup_flush⚠

rocksdb_backup_engine_get_backup_info⚠

rocksdb_backup_engine_info_backup_id⚠

rocksdb_backup_engine_info_count⚠

rocksdb_backup_engine_info_destroy⚠

rocksdb_backup_engine_info_number_files⚠

rocksdb_backup_engine_info_size⚠

rocksdb_backup_engine_info_timestamp⚠

rocksdb_backup_engine_open⚠

rocksdb_backup_engine_open_opts⚠

rocksdb_backup_engine_options_create⚠

rocksdb_backup_engine_options_destroy⚠

rocksdb_backup_engine_options_get_backup_log_files⚠

rocksdb_backup_engine_options_get_backup_rate_limit⚠

rocksdb_backup_engine_options_get_callback_trigger_interval_size⚠

rocksdb_backup_engine_options_get_destroy_old_data⚠

rocksdb_backup_engine_options_get_max_background_operations⚠

rocksdb_backup_engine_options_get_max_valid_backups_to_open⚠

rocksdb_backup_engine_options_get_restore_rate_limit⚠

rocksdb_backup_engine_options_get_share_files_with_checksum_naming⚠

rocksdb_backup_engine_options_get_share_table_files⚠

rocksdb_backup_engine_options_get_sync⚠

rocksdb_backup_engine_options_set_backup_dir⚠

rocksdb_backup_engine_options_set_backup_log_files⚠

rocksdb_backup_engine_options_set_backup_rate_limit⚠

rocksdb_backup_engine_options_set_callback_trigger_interval_size⚠

rocksdb_backup_engine_options_set_destroy_old_data⚠

rocksdb_backup_engine_options_set_env⚠

rocksdb_backup_engine_options_set_max_background_operations⚠

rocksdb_backup_engine_options_set_max_valid_backups_to_open⚠

rocksdb_backup_engine_options_set_restore_rate_limit⚠

rocksdb_backup_engine_options_set_share_files_with_checksum_naming⚠

rocksdb_backup_engine_options_set_share_table_files⚠

rocksdb_backup_engine_options_set_sync⚠

rocksdb_backup_engine_purge_old_backups⚠

rocksdb_backup_engine_restore_db_from_backup⚠

rocksdb_backup_engine_restore_db_from_latest_backup⚠

rocksdb_backup_engine_verify_backup⚠

rocksdb_batched_multi_get_cf⚠

rocksdb_block_based_options_create⚠

rocksdb_block_based_options_destroy⚠

rocksdb_block_based_options_set_block_cache⚠

rocksdb_block_based_options_set_block_restart_interval⚠

rocksdb_block_based_options_set_block_size⚠

rocksdb_block_based_options_set_block_size_deviation⚠

rocksdb_block_based_options_set_cache_index_and_filter_blocks⚠

rocksdb_block_based_options_set_cache_index_and_filter_blocks_with_high_priority⚠

rocksdb_block_based_options_set_checksum⚠

rocksdb_block_based_options_set_data_block_hash_ratio⚠

rocksdb_block_based_options_set_data_block_index_type⚠

rocksdb_block_based_options_set_filter_policy⚠

rocksdb_block_based_options_set_format_version⚠

rocksdb_block_based_options_set_index_block_restart_interval⚠

rocksdb_block_based_options_set_index_type⚠

rocksdb_block_based_options_set_metadata_block_size⚠

rocksdb_block_based_options_set_no_block_cache⚠

rocksdb_block_based_options_set_optimize_filters_for_memory⚠

rocksdb_block_based_options_set_partition_filters⚠

rocksdb_block_based_options_set_pin_l0_filter_and_index_blocks_in_cache⚠

rocksdb_block_based_options_set_pin_top_level_index_and_filter⚠

rocksdb_block_based_options_set_use_delta_encoding⚠

rocksdb_block_based_options_set_whole_key_filtering⚠

rocksdb_cache_create_hyper_clock⚠

rocksdb_cache_create_hyper_clock_opts⚠

rocksdb_cache_create_lru⚠

rocksdb_cache_create_lru_opts⚠

rocksdb_cache_create_lru_with_strict_capacity_limit⚠

rocksdb_cache_destroy⚠

rocksdb_cache_disown_data⚠

rocksdb_cache_get_capacity⚠

rocksdb_cache_get_pinned_usage⚠

rocksdb_cache_get_usage⚠

rocksdb_cache_set_capacity⚠

rocksdb_cancel_all_background_work⚠

rocksdb_checkpoint_create⚠

rocksdb_checkpoint_object_create⚠

rocksdb_checkpoint_object_destroy⚠

rocksdb_close⚠

rocksdb_column_family_handle_destroy⚠

rocksdb_column_family_handle_get_id⚠

rocksdb_column_family_handle_get_name⚠

rocksdb_column_family_metadata_destroy⚠

rocksdb_column_family_metadata_get_file_count⚠

rocksdb_column_family_metadata_get_level_count⚠

rocksdb_column_family_metadata_get_level_metadata⚠

Returns the rocksdb_level_metadata_t of the ith level from the specified column family metadata.

rocksdb_column_family_metadata_get_name⚠

rocksdb_column_family_metadata_get_size⚠

rocksdb_compact_range⚠

rocksdb_compact_range_cf⚠

rocksdb_compact_range_cf_opt⚠

rocksdb_compact_range_opt⚠

rocksdb_compactionfilter_create⚠

rocksdb_compactionfilter_destroy⚠

rocksdb_compactionfilter_set_ignore_snapshots⚠

rocksdb_compactionfiltercontext_is_full_compaction⚠

rocksdb_compactionfiltercontext_is_manual_compaction⚠

rocksdb_compactionfilterfactory_create⚠

rocksdb_compactionfilterfactory_destroy⚠

rocksdb_compactoptions_create⚠

rocksdb_compactoptions_destroy⚠

rocksdb_compactoptions_get_bottommost_level_compaction⚠

rocksdb_compactoptions_get_change_level⚠

rocksdb_compactoptions_get_exclusive_manual_compaction⚠

rocksdb_compactoptions_get_target_level⚠

rocksdb_compactoptions_set_bottommost_level_compaction⚠

rocksdb_compactoptions_set_change_level⚠

rocksdb_compactoptions_set_exclusive_manual_compaction⚠

rocksdb_compactoptions_set_full_history_ts_low⚠

rocksdb_compactoptions_set_target_level⚠

rocksdb_comparator_create⚠

rocksdb_comparator_destroy⚠

rocksdb_comparator_with_ts_create⚠

rocksdb_create_column_family⚠

rocksdb_create_column_family_with_ttl⚠

rocksdb_create_default_env⚠

rocksdb_create_dir_if_missing⚠

rocksdb_create_iterator⚠

rocksdb_create_iterator_cf⚠

rocksdb_create_iterators⚠

rocksdb_create_mem_env⚠

rocksdb_create_snapshot⚠

rocksdb_cuckoo_options_create⚠

rocksdb_cuckoo_options_destroy⚠

rocksdb_cuckoo_options_set_cuckoo_block_size⚠

rocksdb_cuckoo_options_set_hash_ratio⚠

rocksdb_cuckoo_options_set_identity_as_first_hash⚠

rocksdb_cuckoo_options_set_max_search_depth⚠

rocksdb_cuckoo_options_set_use_module_hash⚠

rocksdb_dbpath_create⚠

rocksdb_dbpath_destroy⚠

rocksdb_delete⚠

rocksdb_delete_cf⚠

rocksdb_delete_cf_with_ts⚠

rocksdb_delete_file⚠

rocksdb_delete_file_in_range⚠

rocksdb_delete_file_in_range_cf⚠

rocksdb_delete_range_cf⚠

rocksdb_delete_with_ts⚠

rocksdb_destroy_db⚠

rocksdb_disable_file_deletions⚠

rocksdb_disable_manual_compaction⚠

rocksdb_drop_column_family⚠

rocksdb_enable_file_deletions⚠

rocksdb_enable_manual_compaction⚠

rocksdb_env_destroy⚠

rocksdb_env_get_background_threads⚠

rocksdb_env_get_bottom_priority_background_threads⚠

rocksdb_env_get_high_priority_background_threads⚠

rocksdb_env_get_low_priority_background_threads⚠

rocksdb_env_join_all_threads⚠

rocksdb_env_lower_high_priority_thread_pool_cpu_priority⚠

rocksdb_env_lower_high_priority_thread_pool_io_priority⚠

rocksdb_env_lower_thread_pool_cpu_priority⚠

rocksdb_env_lower_thread_pool_io_priority⚠

rocksdb_env_set_background_threads⚠

rocksdb_env_set_bottom_priority_background_threads⚠

rocksdb_env_set_high_priority_background_threads⚠

rocksdb_env_set_low_priority_background_threads⚠

rocksdb_envoptions_create⚠

rocksdb_envoptions_destroy⚠

rocksdb_fifo_compaction_options_create⚠

rocksdb_fifo_compaction_options_destroy⚠

rocksdb_fifo_compaction_options_get_allow_compaction⚠

rocksdb_fifo_compaction_options_get_max_table_files_size⚠

rocksdb_fifo_compaction_options_set_allow_compaction⚠

rocksdb_fifo_compaction_options_set_max_table_files_size⚠

rocksdb_filterpolicy_create_bloom⚠

rocksdb_filterpolicy_create_bloom_full⚠

rocksdb_filterpolicy_create_ribbon⚠

rocksdb_filterpolicy_create_ribbon_hybrid⚠

rocksdb_filterpolicy_destroy⚠

rocksdb_flush⚠

rocksdb_flush_cf⚠

rocksdb_flush_cfs⚠

rocksdb_flush_wal⚠

rocksdb_flushoptions_create⚠

rocksdb_flushoptions_destroy⚠

rocksdb_flushoptions_get_wait⚠

rocksdb_flushoptions_set_wait⚠

rocksdb_free⚠

rocksdb_get⚠

rocksdb_get_cf⚠

rocksdb_get_cf_with_ts⚠

rocksdb_get_column_family_metadata⚠

rocksdb_get_column_family_metadata_cf⚠

Returns the rocksdb_column_family_metadata_t of the specified column family.

rocksdb_get_full_history_ts_low⚠

rocksdb_get_latest_sequence_number⚠

rocksdb_get_options_from_string⚠

rocksdb_get_pinned⚠

rocksdb_get_pinned_cf⚠

rocksdb_get_updates_since⚠

rocksdb_get_with_ts⚠

rocksdb_hyper_clock_cache_options_create⚠

rocksdb_hyper_clock_cache_options_destroy⚠

rocksdb_hyper_clock_cache_options_set_capacity⚠

rocksdb_hyper_clock_cache_options_set_estimated_entry_charge⚠

rocksdb_hyper_clock_cache_options_set_memory_allocator⚠

rocksdb_hyper_clock_cache_options_set_num_shard_bits⚠

rocksdb_increase_full_history_ts_low⚠

rocksdb_ingest_external_file⚠

rocksdb_ingest_external_file_cf⚠

rocksdb_ingestexternalfileoptions_create⚠

rocksdb_ingestexternalfileoptions_destroy⚠

rocksdb_ingestexternalfileoptions_set_allow_blocking_flush⚠

rocksdb_ingestexternalfileoptions_set_allow_global_seqno⚠

rocksdb_ingestexternalfileoptions_set_fail_if_not_bottommost_level⚠

rocksdb_ingestexternalfileoptions_set_ingest_behind⚠

rocksdb_ingestexternalfileoptions_set_move_files⚠

rocksdb_ingestexternalfileoptions_set_snapshot_consistency⚠

rocksdb_iter_destroy⚠

rocksdb_iter_get_error⚠

rocksdb_iter_key⚠

rocksdb_iter_next⚠

rocksdb_iter_prev⚠

rocksdb_iter_seek⚠

rocksdb_iter_seek_for_prev⚠

rocksdb_iter_seek_to_first⚠

rocksdb_iter_seek_to_last⚠

rocksdb_iter_timestamp⚠

rocksdb_iter_valid⚠

rocksdb_iter_value⚠

rocksdb_jemalloc_nodump_allocator_create⚠

rocksdb_key_may_exist⚠

rocksdb_key_may_exist_cf⚠

rocksdb_level_metadata_destroy⚠

Releases the specified rocksdb_level_metadata_t.

rocksdb_level_metadata_get_file_count⚠

rocksdb_level_metadata_get_level⚠

rocksdb_level_metadata_get_size⚠

rocksdb_level_metadata_get_sst_file_metadata⚠

Returns the sst_file_metadata_t of the ith file from the specified level metadata.

rocksdb_list_column_families⚠

rocksdb_list_column_families_destroy⚠

rocksdb_livefiles⚠

rocksdb_livefiles_column_family_name⚠

rocksdb_livefiles_count⚠

rocksdb_livefiles_deletions⚠

rocksdb_livefiles_destroy⚠

rocksdb_livefiles_entries⚠

rocksdb_livefiles_largestkey⚠

rocksdb_livefiles_level⚠

rocksdb_livefiles_name⚠

rocksdb_livefiles_size⚠

rocksdb_livefiles_smallestkey⚠

rocksdb_load_latest_options⚠

rocksdb_load_latest_options_destroy⚠

rocksdb_lru_cache_options_create⚠

rocksdb_lru_cache_options_destroy⚠

rocksdb_lru_cache_options_set_capacity⚠

rocksdb_lru_cache_options_set_memory_allocator⚠

rocksdb_lru_cache_options_set_num_shard_bits⚠

rocksdb_memory_allocator_destroy⚠

rocksdb_memory_consumers_add_cache⚠

rocksdb_memory_consumers_add_db⚠

rocksdb_memory_consumers_create⚠

rocksdb_memory_consumers_destroy⚠

rocksdb_merge⚠

rocksdb_merge_cf⚠

rocksdb_mergeoperator_create⚠

rocksdb_mergeoperator_destroy⚠

rocksdb_multi_get⚠

rocksdb_multi_get_cf⚠

rocksdb_multi_get_cf_with_ts⚠

rocksdb_multi_get_with_ts⚠

rocksdb_open⚠

rocksdb_open_and_trim_history⚠

rocksdb_open_as_secondary⚠

rocksdb_open_as_secondary_column_families⚠

rocksdb_open_column_families⚠

rocksdb_open_column_families_with_ttl⚠

rocksdb_open_for_read_only⚠

rocksdb_open_for_read_only_column_families⚠

rocksdb_open_with_ttl⚠

rocksdb_optimistictransaction_begin⚠

rocksdb_optimistictransaction_options_create⚠

rocksdb_optimistictransaction_options_destroy⚠

rocksdb_optimistictransaction_options_set_set_snapshot⚠

rocksdb_optimistictransactiondb_checkpoint_object_create⚠

rocksdb_optimistictransactiondb_close⚠

rocksdb_optimistictransactiondb_close_base_db⚠

rocksdb_optimistictransactiondb_get_base_db⚠

rocksdb_optimistictransactiondb_open⚠

rocksdb_optimistictransactiondb_open_column_families⚠

rocksdb_optimistictransactiondb_property_int⚠

rocksdb_optimistictransactiondb_property_value⚠

rocksdb_optimistictransactiondb_write⚠

rocksdb_options_add_compact_on_deletion_collector_factory⚠

rocksdb_options_compaction_readahead_size⚠

rocksdb_options_create⚠

rocksdb_options_create_copy⚠

rocksdb_options_destroy⚠

rocksdb_options_enable_statistics⚠

rocksdb_options_get_WAL_size_limit_MB⚠

rocksdb_options_get_WAL_ttl_seconds⚠

rocksdb_options_get_access_hint_on_compaction_start⚠

rocksdb_options_get_advise_random_on_open⚠

rocksdb_options_get_allow_concurrent_memtable_write⚠

rocksdb_options_get_allow_ingest_behind⚠

rocksdb_options_get_allow_mmap_reads⚠

rocksdb_options_get_allow_mmap_writes⚠

rocksdb_options_get_arena_block_size⚠

rocksdb_options_get_atomic_flush⚠

rocksdb_options_get_avoid_unnecessary_blocking_io⚠

rocksdb_options_get_blob_compaction_readahead_size⚠

rocksdb_options_get_blob_compression_type⚠

rocksdb_options_get_blob_file_size⚠

rocksdb_options_get_blob_file_starting_level⚠

rocksdb_options_get_blob_gc_age_cutoff⚠

rocksdb_options_get_blob_gc_force_threshold⚠

rocksdb_options_get_bloom_locality⚠

rocksdb_options_get_bottommost_compression⚠

rocksdb_options_get_bottommost_compression_options_use_zstd_dict_trainer⚠

rocksdb_options_get_bytes_per_sync⚠

rocksdb_options_get_compaction_readahead_size⚠

rocksdb_options_get_compaction_style⚠

rocksdb_options_get_compression⚠

rocksdb_options_get_compression_options_max_dict_buffer_bytes⚠

rocksdb_options_get_compression_options_parallel_threads⚠

rocksdb_options_get_compression_options_use_zstd_dict_trainer⚠

rocksdb_options_get_compression_options_zstd_max_train_bytes⚠

rocksdb_options_get_create_if_missing⚠

rocksdb_options_get_create_missing_column_families⚠

rocksdb_options_get_db_write_buffer_size⚠

rocksdb_options_get_delete_obsolete_files_period_micros⚠

rocksdb_options_get_disable_auto_compactions⚠

rocksdb_options_get_enable_blob_files⚠

rocksdb_options_get_enable_blob_gc⚠

rocksdb_options_get_enable_pipelined_write⚠

rocksdb_options_get_enable_write_thread_adaptive_yield⚠

rocksdb_options_get_error_if_exists⚠

rocksdb_options_get_experimental_mempurge_threshold⚠

rocksdb_options_get_hard_pending_compaction_bytes_limit⚠

rocksdb_options_get_info_log_level⚠

rocksdb_options_get_inplace_update_num_locks⚠

rocksdb_options_get_inplace_update_support⚠

rocksdb_options_get_is_fd_close_on_exec⚠

rocksdb_options_get_keep_log_file_num⚠

rocksdb_options_get_level0_file_num_compaction_trigger⚠

rocksdb_options_get_level0_slowdown_writes_trigger⚠

rocksdb_options_get_level0_stop_writes_trigger⚠

rocksdb_options_get_level_compaction_dynamic_level_bytes⚠

rocksdb_options_get_log_file_time_to_roll⚠

rocksdb_options_get_manifest_preallocation_size⚠

rocksdb_options_get_manual_wal_flush⚠

rocksdb_options_get_max_background_compactions⚠

rocksdb_options_get_max_background_flushes⚠

rocksdb_options_get_max_background_jobs⚠

rocksdb_options_get_max_bytes_for_level_base⚠

rocksdb_options_get_max_bytes_for_level_multiplier⚠

rocksdb_options_get_max_compaction_bytes⚠

rocksdb_options_get_max_file_opening_threads⚠

rocksdb_options_get_max_log_file_size⚠

rocksdb_options_get_max_manifest_file_size⚠

rocksdb_options_get_max_open_files⚠

rocksdb_options_get_max_sequential_skip_in_iterations⚠

rocksdb_options_get_max_subcompactions⚠

rocksdb_options_get_max_successive_merges⚠

rocksdb_options_get_max_total_wal_size⚠

rocksdb_options_get_max_write_buffer_number⚠

rocksdb_options_get_max_write_buffer_number_to_maintain⚠

rocksdb_options_get_max_write_buffer_size_to_maintain⚠

rocksdb_options_get_memtable_huge_page_size⚠

rocksdb_options_get_memtable_prefix_bloom_size_ratio⚠

rocksdb_options_get_min_blob_size⚠

rocksdb_options_get_min_write_buffer_number_to_merge⚠

rocksdb_options_get_num_levels⚠

rocksdb_options_get_optimize_filters_for_hits⚠

rocksdb_options_get_paranoid_checks⚠

rocksdb_options_get_prepopulate_blob_cache⚠

rocksdb_options_get_recycle_log_file_num⚠

rocksdb_options_get_report_bg_io_stats⚠

rocksdb_options_get_skip_checking_sst_file_sizes_on_db_open⚠

rocksdb_options_get_skip_stats_update_on_db_open⚠

rocksdb_options_get_soft_pending_compaction_bytes_limit⚠

rocksdb_options_get_stats_dump_period_sec⚠

rocksdb_options_get_stats_persist_period_sec⚠

rocksdb_options_get_table_cache_numshardbits⚠

rocksdb_options_get_target_file_size_base⚠

rocksdb_options_get_target_file_size_multiplier⚠

rocksdb_options_get_unordered_write⚠

rocksdb_options_get_use_adaptive_mutex⚠

rocksdb_options_get_use_direct_io_for_flush_and_compaction⚠

rocksdb_options_get_use_direct_reads⚠

rocksdb_options_get_use_fsync⚠

rocksdb_options_get_wal_bytes_per_sync⚠

rocksdb_options_get_wal_compression⚠

rocksdb_options_get_wal_recovery_mode⚠

rocksdb_options_get_writable_file_max_buffer_size⚠

rocksdb_options_get_write_buffer_size⚠

rocksdb_options_increase_parallelism⚠

rocksdb_options_optimize_for_point_lookup⚠

rocksdb_options_optimize_level_style_compaction⚠

rocksdb_options_optimize_universal_style_compaction⚠

rocksdb_options_prepare_for_bulk_load⚠

rocksdb_options_set_WAL_size_limit_MB⚠

rocksdb_options_set_WAL_ttl_seconds⚠

rocksdb_options_set_access_hint_on_compaction_start⚠

rocksdb_options_set_advise_random_on_open⚠

rocksdb_options_set_allow_concurrent_memtable_write⚠

rocksdb_options_set_allow_ingest_behind⚠

rocksdb_options_set_allow_mmap_reads⚠

rocksdb_options_set_allow_mmap_writes⚠

rocksdb_options_set_arena_block_size⚠

rocksdb_options_set_atomic_flush⚠

rocksdb_options_set_avoid_unnecessary_blocking_io⚠

rocksdb_options_set_blob_cache⚠

rocksdb_options_set_blob_compaction_readahead_size⚠

rocksdb_options_set_blob_compression_type⚠

rocksdb_options_set_blob_file_size⚠

rocksdb_options_set_blob_file_starting_level⚠

rocksdb_options_set_blob_gc_age_cutoff⚠

rocksdb_options_set_blob_gc_force_threshold⚠

rocksdb_options_set_block_based_table_factory⚠

rocksdb_options_set_bloom_locality⚠

rocksdb_options_set_bottommost_compression⚠

rocksdb_options_set_bottommost_compression_options⚠

rocksdb_options_set_bottommost_compression_options_max_dict_buffer_bytes⚠

rocksdb_options_set_bottommost_compression_options_use_zstd_dict_trainer⚠

rocksdb_options_set_bottommost_compression_options_zstd_max_train_bytes⚠

rocksdb_options_set_bytes_per_sync⚠

rocksdb_options_set_compaction_filter⚠

rocksdb_options_set_compaction_filter_factory⚠

rocksdb_options_set_compaction_style⚠

rocksdb_options_set_comparator⚠

rocksdb_options_set_compression⚠

rocksdb_options_set_compression_options⚠

rocksdb_options_set_compression_options_max_dict_buffer_bytes⚠

rocksdb_options_set_compression_options_parallel_threads⚠

rocksdb_options_set_compression_options_use_zstd_dict_trainer⚠

rocksdb_options_set_compression_options_zstd_max_train_bytes⚠

rocksdb_options_set_compression_per_level⚠

rocksdb_options_set_create_if_missing⚠

rocksdb_options_set_create_missing_column_families⚠

rocksdb_options_set_cuckoo_table_factory⚠

rocksdb_options_set_db_log_dir⚠

rocksdb_options_set_db_paths⚠

rocksdb_options_set_db_write_buffer_size⚠

rocksdb_options_set_delete_obsolete_files_period_micros⚠

rocksdb_options_set_disable_auto_compactions⚠

rocksdb_options_set_dump_malloc_stats⚠

rocksdb_options_set_enable_blob_files⚠

rocksdb_options_set_enable_blob_gc⚠

rocksdb_options_set_enable_pipelined_write⚠

rocksdb_options_set_enable_write_thread_adaptive_yield⚠

rocksdb_options_set_env⚠

rocksdb_options_set_error_if_exists⚠

rocksdb_options_set_experimental_mempurge_threshold⚠

rocksdb_options_set_fifo_compaction_options⚠

rocksdb_options_set_hard_pending_compaction_bytes_limit⚠

rocksdb_options_set_hash_link_list_rep⚠

rocksdb_options_set_hash_skip_list_rep⚠

rocksdb_options_set_info_log⚠

rocksdb_options_set_info_log_level⚠

rocksdb_options_set_inplace_update_num_locks⚠

rocksdb_options_set_inplace_update_support⚠

rocksdb_options_set_is_fd_close_on_exec⚠

rocksdb_options_set_keep_log_file_num⚠

rocksdb_options_set_level0_file_num_compaction_trigger⚠

rocksdb_options_set_level0_slowdown_writes_trigger⚠

rocksdb_options_set_level0_stop_writes_trigger⚠

rocksdb_options_set_level_compaction_dynamic_level_bytes⚠

rocksdb_options_set_log_file_time_to_roll⚠

rocksdb_options_set_manifest_preallocation_size⚠

rocksdb_options_set_manual_wal_flush⚠

rocksdb_options_set_max_background_compactions⚠

rocksdb_options_set_max_background_flushes⚠

rocksdb_options_set_max_background_jobs⚠

rocksdb_options_set_max_bytes_for_level_base⚠

rocksdb_options_set_max_bytes_for_level_multiplier⚠

rocksdb_options_set_max_bytes_for_level_multiplier_additional⚠

rocksdb_options_set_max_compaction_bytes⚠

rocksdb_options_set_max_file_opening_threads⚠

rocksdb_options_set_max_log_file_size⚠

rocksdb_options_set_max_manifest_file_size⚠

rocksdb_options_set_max_open_files⚠

rocksdb_options_set_max_sequential_skip_in_iterations⚠

rocksdb_options_set_max_subcompactions⚠

rocksdb_options_set_max_successive_merges⚠

rocksdb_options_set_max_total_wal_size⚠

rocksdb_options_set_max_write_buffer_number⚠

rocksdb_options_set_max_write_buffer_number_to_maintain⚠

rocksdb_options_set_max_write_buffer_size_to_maintain⚠

rocksdb_options_set_memtable_huge_page_size⚠

rocksdb_options_set_memtable_prefix_bloom_size_ratio⚠

rocksdb_options_set_memtable_vector_rep⚠

rocksdb_options_set_memtable_whole_key_filtering⚠

rocksdb_options_set_merge_operator⚠

rocksdb_options_set_min_blob_size⚠

rocksdb_options_set_min_level_to_compress⚠

rocksdb_options_set_min_write_buffer_number_to_merge⚠

rocksdb_options_set_num_levels⚠

rocksdb_options_set_optimize_filters_for_hits⚠

rocksdb_options_set_paranoid_checks⚠

rocksdb_options_set_plain_table_factory⚠

rocksdb_options_set_prefix_extractor⚠

rocksdb_options_set_prepopulate_blob_cache⚠

rocksdb_options_set_ratelimiter⚠

rocksdb_options_set_recycle_log_file_num⚠

rocksdb_options_set_report_bg_io_stats⚠

rocksdb_options_set_row_cache⚠

rocksdb_options_set_skip_checking_sst_file_sizes_on_db_open⚠

rocksdb_options_set_skip_stats_update_on_db_open⚠

rocksdb_options_set_soft_pending_compaction_bytes_limit⚠

rocksdb_options_set_stats_dump_period_sec⚠

rocksdb_options_set_stats_persist_period_sec⚠

rocksdb_options_set_table_cache_numshardbits⚠

rocksdb_options_set_target_file_size_base⚠

rocksdb_options_set_target_file_size_multiplier⚠

rocksdb_options_set_uint64add_merge_operator⚠

rocksdb_options_set_universal_compaction_options⚠

rocksdb_options_set_unordered_write⚠

rocksdb_options_set_use_adaptive_mutex⚠

rocksdb_options_set_use_direct_io_for_flush_and_compaction⚠

rocksdb_options_set_use_direct_reads⚠

rocksdb_options_set_use_fsync⚠

rocksdb_options_set_wal_bytes_per_sync⚠

rocksdb_options_set_wal_compression⚠

rocksdb_options_set_wal_dir⚠

rocksdb_options_set_wal_recovery_mode⚠

rocksdb_options_set_writable_file_max_buffer_size⚠

rocksdb_options_set_write_buffer_size⚠

rocksdb_options_statistics_get_string⚠

rocksdb_perfcontext_create⚠

rocksdb_perfcontext_destroy⚠

rocksdb_perfcontext_metric⚠

rocksdb_perfcontext_report⚠

rocksdb_perfcontext_reset⚠

rocksdb_pinnableslice_destroy⚠

rocksdb_pinnableslice_value⚠

rocksdb_property_int⚠

rocksdb_property_int_cf⚠

rocksdb_property_value⚠

rocksdb_property_value_cf⚠

rocksdb_put⚠

rocksdb_put_cf⚠

rocksdb_put_cf_with_ts⚠

rocksdb_put_with_ts⚠

rocksdb_ratelimiter_create⚠

rocksdb_ratelimiter_destroy⚠

rocksdb_readoptions_create⚠

rocksdb_readoptions_destroy⚠

rocksdb_readoptions_get_async_io⚠

rocksdb_readoptions_get_background_purge_on_iterator_cleanup⚠

rocksdb_readoptions_get_deadline⚠

rocksdb_readoptions_get_fill_cache⚠

rocksdb_readoptions_get_ignore_range_deletions⚠

rocksdb_readoptions_get_io_timeout⚠

rocksdb_readoptions_get_max_skippable_internal_keys⚠

rocksdb_readoptions_get_pin_data⚠

rocksdb_readoptions_get_prefix_same_as_start⚠

rocksdb_readoptions_get_read_tier⚠

rocksdb_readoptions_get_readahead_size⚠

rocksdb_readoptions_get_tailing⚠

rocksdb_readoptions_get_total_order_seek⚠

rocksdb_readoptions_get_verify_checksums⚠

rocksdb_readoptions_set_async_io⚠

rocksdb_readoptions_set_background_purge_on_iterator_cleanup⚠

rocksdb_readoptions_set_deadline⚠

rocksdb_readoptions_set_fill_cache⚠

rocksdb_readoptions_set_ignore_range_deletions⚠

rocksdb_readoptions_set_io_timeout⚠

rocksdb_readoptions_set_iter_start_ts⚠

rocksdb_readoptions_set_iterate_lower_bound⚠

rocksdb_readoptions_set_iterate_upper_bound⚠

rocksdb_readoptions_set_managed⚠

rocksdb_readoptions_set_max_skippable_internal_keys⚠

rocksdb_readoptions_set_pin_data⚠

rocksdb_readoptions_set_prefix_same_as_start⚠

rocksdb_readoptions_set_read_tier⚠

rocksdb_readoptions_set_readahead_size⚠

rocksdb_readoptions_set_snapshot⚠

rocksdb_readoptions_set_tailing⚠

rocksdb_readoptions_set_timestamp⚠

rocksdb_readoptions_set_total_order_seek⚠

rocksdb_readoptions_set_verify_checksums⚠

rocksdb_release_snapshot⚠

rocksdb_repair_db⚠

rocksdb_restore_options_create⚠

rocksdb_restore_options_destroy⚠

rocksdb_restore_options_set_keep_log_files⚠

rocksdb_set_options⚠

rocksdb_set_options_cf⚠

rocksdb_set_perf_level⚠

rocksdb_singledelete⚠

rocksdb_singledelete_cf⚠

rocksdb_singledelete_cf_with_ts⚠

rocksdb_singledelete_with_ts⚠

rocksdb_slicetransform_create⚠

rocksdb_slicetransform_create_fixed_prefix⚠

rocksdb_slicetransform_create_noop⚠

rocksdb_slicetransform_destroy⚠

rocksdb_sst_file_metadata_destroy⚠

Releases the specified rocksdb_sst_file_metadata_t.

rocksdb_sst_file_metadata_get_largestkey⚠

Returns the smallest key of the specified sst file. The caller is responsible for releasing the returned memory.

rocksdb_sst_file_metadata_get_relative_filename⚠

rocksdb_sst_file_metadata_get_size⚠

rocksdb_sst_file_metadata_get_smallestkey⚠

Returns the smallest key of the specified sst file. The caller is responsible for releasing the returned memory.

rocksdb_sstfilewriter_add⚠

rocksdb_sstfilewriter_create⚠

rocksdb_sstfilewriter_create_with_comparator⚠

rocksdb_sstfilewriter_delete⚠

rocksdb_sstfilewriter_delete_range⚠

rocksdb_sstfilewriter_delete_with_ts⚠

rocksdb_sstfilewriter_destroy⚠

rocksdb_sstfilewriter_file_size⚠

rocksdb_sstfilewriter_finish⚠

rocksdb_sstfilewriter_merge⚠

rocksdb_sstfilewriter_open⚠

rocksdb_sstfilewriter_put⚠

rocksdb_sstfilewriter_put_with_ts⚠

rocksdb_suggest_compact_range⚠

rocksdb_suggest_compact_range_cf⚠

rocksdb_transaction_begin⚠

rocksdb_transaction_commit⚠

rocksdb_transaction_create_iterator⚠

rocksdb_transaction_create_iterator_cf⚠

rocksdb_transaction_delete⚠

rocksdb_transaction_delete_cf⚠

rocksdb_transaction_destroy⚠

rocksdb_transaction_get⚠

rocksdb_transaction_get_cf⚠

rocksdb_transaction_get_for_update⚠

rocksdb_transaction_get_for_update_cf⚠

rocksdb_transaction_get_name⚠

rocksdb_transaction_get_pinned⚠

rocksdb_transaction_get_pinned_cf⚠

rocksdb_transaction_get_pinned_for_update⚠

rocksdb_transaction_get_pinned_for_update_cf⚠

rocksdb_transaction_get_snapshot⚠

rocksdb_transaction_get_writebatch_wi⚠

rocksdb_transaction_merge⚠

rocksdb_transaction_merge_cf⚠

rocksdb_transaction_multi_get⚠

rocksdb_transaction_multi_get_cf⚠

rocksdb_transaction_options_create⚠

rocksdb_transaction_options_destroy⚠

rocksdb_transaction_options_set_deadlock_detect⚠

rocksdb_transaction_options_set_deadlock_detect_depth⚠

rocksdb_transaction_options_set_expiration⚠

rocksdb_transaction_options_set_lock_timeout⚠

rocksdb_transaction_options_set_max_write_batch_size⚠

rocksdb_transaction_options_set_set_snapshot⚠

rocksdb_transaction_options_set_skip_prepare⚠

rocksdb_transaction_prepare⚠

rocksdb_transaction_put⚠

rocksdb_transaction_put_cf⚠

rocksdb_transaction_rebuild_from_writebatch⚠

rocksdb_transaction_rebuild_from_writebatch_wi⚠

rocksdb_transaction_rollback⚠

rocksdb_transaction_rollback_to_savepoint⚠

rocksdb_transaction_set_commit_timestamp⚠

rocksdb_transaction_set_name⚠

rocksdb_transaction_set_read_timestamp_for_validation⚠

rocksdb_transaction_set_savepoint⚠

rocksdb_transactiondb_checkpoint_object_create⚠

rocksdb_transactiondb_close⚠

rocksdb_transactiondb_create_column_family⚠

rocksdb_transactiondb_create_iterator⚠

rocksdb_transactiondb_create_iterator_cf⚠

rocksdb_transactiondb_create_snapshot⚠

rocksdb_transactiondb_delete⚠

rocksdb_transactiondb_delete_cf⚠

rocksdb_transactiondb_flush⚠

rocksdb_transactiondb_flush_cf⚠

rocksdb_transactiondb_flush_cfs⚠

rocksdb_transactiondb_flush_wal⚠

rocksdb_transactiondb_get⚠

rocksdb_transactiondb_get_cf⚠

rocksdb_transactiondb_get_pinned⚠

rocksdb_transactiondb_get_pinned_cf⚠

rocksdb_transactiondb_get_prepared_transactions⚠

rocksdb_transactiondb_merge⚠

rocksdb_transactiondb_merge_cf⚠

rocksdb_transactiondb_multi_get⚠

rocksdb_transactiondb_multi_get_cf⚠

rocksdb_transactiondb_open⚠

rocksdb_transactiondb_open_column_families⚠

rocksdb_transactiondb_options_create⚠

rocksdb_transactiondb_options_destroy⚠

rocksdb_transactiondb_options_set_default_lock_timeout⚠

rocksdb_transactiondb_options_set_max_num_locks⚠

rocksdb_transactiondb_options_set_num_stripes⚠

rocksdb_transactiondb_options_set_transaction_lock_timeout⚠

rocksdb_transactiondb_property_int⚠

rocksdb_transactiondb_property_value⚠

rocksdb_transactiondb_put⚠

rocksdb_transactiondb_put_cf⚠

rocksdb_transactiondb_release_snapshot⚠

rocksdb_transactiondb_write⚠

rocksdb_try_catch_up_with_primary⚠

rocksdb_universal_compaction_options_create⚠

rocksdb_universal_compaction_options_destroy⚠

rocksdb_universal_compaction_options_get_compression_size_percent⚠

rocksdb_universal_compaction_options_get_max_merge_width⚠

rocksdb_universal_compaction_options_get_max_size_amplification_percent⚠

rocksdb_universal_compaction_options_get_min_merge_width⚠

rocksdb_universal_compaction_options_get_size_ratio⚠

rocksdb_universal_compaction_options_get_stop_style⚠

rocksdb_universal_compaction_options_set_compression_size_percent⚠

rocksdb_universal_compaction_options_set_max_merge_width⚠

rocksdb_universal_compaction_options_set_max_size_amplification_percent⚠

rocksdb_universal_compaction_options_set_min_merge_width⚠

rocksdb_universal_compaction_options_set_size_ratio⚠

rocksdb_universal_compaction_options_set_stop_style⚠

rocksdb_wal_iter_destroy⚠

rocksdb_wal_iter_get_batch⚠

rocksdb_wal_iter_next⚠

rocksdb_wal_iter_status⚠

rocksdb_wal_iter_valid⚠

rocksdb_write⚠

rocksdb_write_writebatch_wi⚠

rocksdb_writebatch_clear⚠

rocksdb_writebatch_count⚠

rocksdb_writebatch_create⚠

rocksdb_writebatch_create_from⚠

rocksdb_writebatch_data⚠

rocksdb_writebatch_delete⚠

rocksdb_writebatch_delete_cf⚠

rocksdb_writebatch_delete_cf_with_ts⚠

rocksdb_writebatch_delete_range⚠

rocksdb_writebatch_delete_range_cf⚠

rocksdb_writebatch_delete_rangev⚠

rocksdb_writebatch_delete_rangev_cf⚠

rocksdb_writebatch_deletev⚠

rocksdb_writebatch_deletev_cf⚠

rocksdb_writebatch_destroy⚠

rocksdb_writebatch_iterate⚠

rocksdb_writebatch_merge⚠

rocksdb_writebatch_merge_cf⚠

rocksdb_writebatch_mergev⚠

rocksdb_writebatch_mergev_cf⚠

rocksdb_writebatch_pop_save_point⚠

rocksdb_writebatch_put⚠

rocksdb_writebatch_put_cf⚠

rocksdb_writebatch_put_cf_with_ts⚠

rocksdb_writebatch_put_log_data⚠

rocksdb_writebatch_putv⚠

rocksdb_writebatch_putv_cf⚠

rocksdb_writebatch_rollback_to_save_point⚠

rocksdb_writebatch_set_save_point⚠

rocksdb_writebatch_singledelete⚠

rocksdb_writebatch_singledelete_cf⚠

rocksdb_writebatch_singledelete_cf_with_ts⚠

rocksdb_writebatch_wi_clear⚠

rocksdb_writebatch_wi_count⚠

rocksdb_writebatch_wi_create⚠

rocksdb_writebatch_wi_create_from⚠

rocksdb_writebatch_wi_create_iterator_with_base⚠

rocksdb_writebatch_wi_create_iterator_with_base_cf⚠

rocksdb_writebatch_wi_data⚠

rocksdb_writebatch_wi_delete⚠

rocksdb_writebatch_wi_delete_cf⚠

rocksdb_writebatch_wi_delete_range⚠

rocksdb_writebatch_wi_delete_range_cf⚠

rocksdb_writebatch_wi_delete_rangev⚠

rocksdb_writebatch_wi_delete_rangev_cf⚠

rocksdb_writebatch_wi_deletev⚠

rocksdb_writebatch_wi_deletev_cf⚠

rocksdb_writebatch_wi_destroy⚠

rocksdb_writebatch_wi_get_from_batch⚠

rocksdb_writebatch_wi_get_from_batch_and_db⚠

rocksdb_writebatch_wi_get_from_batch_and_db_cf⚠

rocksdb_writebatch_wi_get_from_batch_cf⚠

rocksdb_writebatch_wi_iterate⚠

rocksdb_writebatch_wi_merge⚠

rocksdb_writebatch_wi_merge_cf⚠

rocksdb_writebatch_wi_mergev⚠

rocksdb_writebatch_wi_mergev_cf⚠

rocksdb_writebatch_wi_put⚠

rocksdb_writebatch_wi_put_cf⚠

rocksdb_writebatch_wi_put_log_data⚠

rocksdb_writebatch_wi_putv⚠

rocksdb_writebatch_wi_putv_cf⚠

rocksdb_writebatch_wi_rollback_to_save_point⚠

rocksdb_writebatch_wi_set_save_point⚠

rocksdb_writebatch_wi_singledelete⚠

rocksdb_writebatch_wi_singledelete_cf⚠

rocksdb_writeoptions_create⚠

rocksdb_writeoptions_destroy⚠

rocksdb_writeoptions_disable_WAL⚠

rocksdb_writeoptions_get_disable_WAL⚠

rocksdb_writeoptions_get_ignore_missing_column_families⚠

rocksdb_writeoptions_get_low_pri⚠

rocksdb_writeoptions_get_memtable_insert_hint_per_batch⚠

rocksdb_writeoptions_get_no_slowdown⚠

rocksdb_writeoptions_get_sync⚠

rocksdb_writeoptions_set_ignore_missing_column_families⚠

rocksdb_writeoptions_set_low_pri⚠

rocksdb_writeoptions_set_memtable_insert_hint_per_batch⚠

rocksdb_writeoptions_set_no_slowdown⚠

rocksdb_writeoptions_set_sync⚠

Type Aliases

__blkcnt64_t

__blkcnt_t

__blksize_t

__builtin_va_list

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__gnuc_va_list

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

_bindgen_ty_1

_bindgen_ty_2

_bindgen_ty_3

_bindgen_ty_4

_bindgen_ty_5

_bindgen_ty_6

_bindgen_ty_7

_bindgen_ty_8

_bindgen_ty_9

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

va_list

wchar_t